+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Enhancements
 
+ * `bindings` is a new command group for operations on bindings
  * `exchanges` is a new command group for operations on exchanges
  * `global_parameters` is a new command group for operations on [global runtime parameters](https://www.rabbitmq.com/docs/parameters)
  * `nodes` is a new command group for operations on nodes